// appointza-ui-canvas/src/models/users.model.ts
export class UsersRegisterReq {
  organisationimageid: number = 0;
  organisationname: string = '';
  organisationgstnumber: string = '';
  organisationlogo: number = 0;
  secondarytypecode: string = ""
  secondarytype: number = 0
  primarytype: number = 0
  primarytypecode: string = ""
  organisationtype: number = 0;
  latitude: number = 0;
  longitude: number = 0;
  googlelocation: string = '';
  locationname: string = '';
  locationaddressline1: string = '';
  locationaddressline2: string = '';
  locationcity: string = '';
  locationstate: string = '';
  locationcountry: string = '';
  locationpincode: string = '';
  username: string = '';
  useremail: string = '';
  usermobile: string = '';
  usermobilecountrycode: string = '';
  userdesignation: string = '';
  profileimage: number = 0;
  plan_code: string = '';
  referral_code: string = '';
}

export class UsersRegisterRes {
  mobile: string = '';
  name: string = '';
}

export class UsersLoginReq {
  mobile: string = '';
  otp: string = '';
}
// ...
export class UsersGetOtpReq {
  mobile: string = '';
}